Added financial controller

+++ src/main/java/com/daduo/api/tiktokapi/model/ExchangeRequest.java
@@ -11,7 +11,7 @@ public class ExchangeRequest {
     @ApiModelProperty(value = "用户ID")
     private Long userId;
     @ApiModelProperty(value = "兑换金额")
-    private Long money;
+    private Integer money;
     @ApiModelProperty(value = "兑换方式")
     private ExchangeMethod exchangeMethod;
     @ApiModelProperty(value = "支付宝或者微信二维码地址")
